Fixes bug in getGroupsForUser related to periods
When using roles that include periods (`.`), as an example in the docs, `getGroupsForUser` will always return an empty array is one of those roles is passed.

Group names are converted by replacing periods (`.`) with underscores (`_`) which is needed as groups names are used as roles object keys. Role names are never converted this way so the commented line seems to make non sense here.